03 Python/Entwicklung eines IoT-Devices/loesung/requirements.txt
apsw            # SQLite-Wrapper
cherrypy        # Eingebetteter Webserver
Adafruit-DHT    # DHT11-Sensor
RPi.GPIO        # Raspberry Pi GPIO
